fix imports/structure so things work now, kinda
